Changed readme.txt to fit wordpress style, make sure [[curl]] is used
--HG--
extra : convert_revision : svn%3Ab8457f37-d9ea-0310-8a92-e5e31aec5664/wordpress-petition-plugin/trunk%4025144

+++ fcpetition.php
@@ -356,13 +356,27 @@ function fcpetition_options_page() {
             // Save the posted value in the database
 	    update_option("petition_title", $petition_title );
             update_option("petition_text", $petition_text );
-	    update_option("petition_confirmation", $petition_confirmation );
+	    if(strpos($petition_confirmation,"[[curl]]")) {
+	    	update_option("petition_confirmation", $petition_confirmation );
+	    } else {
+		$p_error = __("[[curl]] must appear in your confirmation email text.","fcpetition");
+		$petition_confirmation = get_option("petition_confirmation");
+	    }
 	    update_option("petition_confirmurl", $petition_confirmurl );
 	    update_option("petition_from", $petition_from );
 	    update_option("petition_maximum", $petition_maximum );
 	    update_option("petition_enabled", $petition_enabled );
 	    update_option("petition_comments", $petition_comments );
             // Put an options updated message on the sc
+
+	    if($p_error != "") {
+		print "
+			<div id=\"message\" class=\"error fade\"><p><strong>
+				$p_error
+	                </p></strong></div>
+		";
+	    }
+
 	    ?>
 	    <div id="message" class="updated fade"><p><strong>
 		    <?php _e("Options Updated.","fcpetition") ?>
